move is_nonstr_iter to pyramid.util
