+++ src/pyramid/authorization.py
@@ -4,9 +4,9 @@ from pyramid.interfaces import IAuthorizationPolicy
 
 from pyramid.location import lineage
 
-from pyramid.compat import is_nonstr_iter
-
 from pyramid.security import ACLAllowed, ACLDenied, Allow, Deny, Everyone
+
+from pyramid.util import is_nonstr_iter
 
 
 @implementer(IAuthorizationPolicy)
